Move style-related utility functions to styleUtils.js
