+++ docs/assets/sldreader.js
@@ -822,45 +822,6 @@
   }
 
   /**
-   * Function to memoize style conversion functions that convert sld symbolizers to OpenLayers style instances.
-   * The memoized version of the style converter returns the same OL style instance if the symbolizer is the same object.
-   * Uses a WeakMap internally.
-   * Note: This only works for constant symbolizers.
-   * @private
-   * @param {Function} styleFunction Function that accepts a single symbolizer object and returns the corresponding OpenLayers style object.
-   * @returns {Function} The memoized function of the style conversion function.
-   */
-  function memoizeStyleFunction(styleFunction) {
-    var styleCache = new WeakMap();
-
-    return function (symbolizer) {
-      var olStyle = styleCache.get(symbolizer);
-
-      if (!olStyle) {
-        olStyle = styleFunction(symbolizer);
-        styleCache.set(symbolizer, olStyle);
-      }
-
-      return olStyle;
-    };
-  }
-
-  /**
-   * @param  {string} hex   eg #AA00FF
-   * @param  {Number} alpha eg 0.5
-   * @return {string}       rgba(0,0,0,0)
-   */
-  function hexToRGB(hex, alpha) {
-    var r = parseInt(hex.slice(1, 3), 16);
-    var g = parseInt(hex.slice(3, 5), 16);
-    var b = parseInt(hex.slice(5, 7), 16);
-    if (alpha) {
-      return ("rgba(" + r + ", " + g + ", " + b + ", " + alpha + ")");
-    }
-    return ("rgb(" + r + ", " + g + ", " + b + ")");
-  }
-
-  /**
    * Get styling from rules per geometry type
    * @param  {Rule[]} rules [description]
    * @return {GeometryStyles}
@@ -1194,6 +1155,45 @@
       width: 1,
     }),
   });
+
+  /**
+   * Function to memoize style conversion functions that convert sld symbolizers to OpenLayers style instances.
+   * The memoized version of the style converter returns the same OL style instance if the symbolizer is the same object.
+   * Uses a WeakMap internally.
+   * Note: This only works for constant symbolizers.
+   * @private
+   * @param {Function} styleFunction Function that accepts a single symbolizer object and returns the corresponding OpenLayers style object.
+   * @returns {Function} The memoized function of the style conversion function.
+   */
+  function memoizeStyleFunction(styleFunction) {
+    var styleCache = new WeakMap();
+
+    return function (symbolizer) {
+      var olStyle = styleCache.get(symbolizer);
+
+      if (!olStyle) {
+        olStyle = styleFunction(symbolizer);
+        styleCache.set(symbolizer, olStyle);
+      }
+
+      return olStyle;
+    };
+  }
+
+  /**
+   * @param  {string} hex   eg #AA00FF
+   * @param  {Number} alpha eg 0.5
+   * @return {string}       rgba(0,0,0,0)
+   */
+  function hexToRGB(hex, alpha) {
+    var r = parseInt(hex.slice(1, 3), 16);
+    var g = parseInt(hex.slice(3, 5), 16);
+    var b = parseInt(hex.slice(5, 7), 16);
+    if (alpha) {
+      return ("rgba(" + r + ", " + g + ", " + b + ", " + alpha + ")");
+    }
+    return ("rgb(" + r + ", " + g + ", " + b + ")");
+  }
 
   /**
    * Create an OL point style corresponding to a well known symbol identifier.
@@ -1931,8 +1931,6 @@
   exports.getRules = getRules;
   exports.getStyle = getStyle;
   exports.getStyleNames = getStyleNames;
-  exports.hexToRGB = hexToRGB;
-  exports.memoizeStyleFunction = memoizeStyleFunction;
 
   Object.defineProperty(exports, '__esModule', { value: true });
